Moved functions in CssGen that produce about page and timetable CSS to respective files.
